Adds Unsubscribe method and ticket system for subscribers
First of all we add support for ticketing on subscriber store.
This will be needed when connections ends. Some refactors
done to add this functionality. Also added one built in ticket generator
based on standard lib. We are trying to prevent importing external libs.
The Close() method now is Reset() method. It does a complete clean up of
the store by closing channels and de referencing all subscribers.
Once reset, the store can be reused.